requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
environs==10.0.0
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
packaging==23.2
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
SQLAlchemy==1.3.24
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
